requirements.txt
Django==5.2.1
python-dotenv==1.0.1
mysqlclient==2.2.7
requests==2.32.3
openpyxl==3.1.5
pandas==2.2.3
psycopg==3.2.9
